move ChoroplethLegend into compositions
