+++ src/ui/compositions/choropleth-legend/README.md
