+++ src/ui/compositions/choropleth-legend/_readme.template.md
