+++ src/ui/compositions/choropleth-legend/demo/index.html
