+++ src/ui/compositions/choropleth-legend/index.js
