+++ src/ui/compositions/choropleth-legend/src/choropleth-legend.css
