+++ src/ui/compositions/choropleth-legend/src/linear-gradient.jsx
