+++ src/ui/compositions/choropleth-legend/src/slider-handle.css
