+++ src/ui/compositions/choropleth-legend/test/.eslintrc.js
