+++ src/ui/compositions/choropleth-legend/test/slider-handle.test.js
