+++ src/ui/compositions/choropleth-legend/test/slider.test.js
